esp32/modnetwork: Add interface to the WIFI power save functions of ESP-IDF.
New function WLAN.wifi_ps() with zero or one parameter.
New constants in network:
  WIFI_PS_NONE
  WIFI_PS_MIN_MODEM
  WIFI_PS_MAX_MODEM

When called without parameter, the function returns the current WIFI power save mode.
When called with one parameter, the power save mode is set to the value of the parameter and the previous power save mode is returned.

esp32/modnetwork: code formatting